% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allcas}
\alias{allcas}
\title{cas list for N sample>=1 per year and Nchemicals>=1 per year .}
\format{
A data frame with 334 rows and 18 variables:
\describe{
\item{CAS}{test id}
\item{Chem.Group}{Industrial, Metal, PAH, Pest, Pharma}
...
}
}
\source{
\url{https://cfpub.epa.gov/ecotox/}
}
\usage{
allcas
}
\description{
A dataset containing the chemical classification data
}
\keyword{datasets}
