% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_LOQ.R
\name{annotation_custom2}
\alias{annotation_custom2}
\title{layerd plot definitions in inset_plot}
\usage{
annotation_custom2(
  grob,
  xmin = -Inf,
  xmax = Inf,
  ymin = -Inf,
  ymax = Inf,
  data
)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
layerd plot definitions in inset_plot
}
