% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chemclass}
\alias{chemclass}
\title{Dataset for Chemical classification.}
\format{
A data frame with 334 rows and 18 variables:
\describe{
\item{CAS}{test id}
\item{Chem.Group}{Industrial, Metal, PAH, Pest, Pharma}
...
}
}
\source{
Check with repos owner
}
\usage{
chemclass
}
\description{
A dataset containing the chemical classification data
}
\keyword{datasets}
