% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data.AggBySiteID.2}
\alias{Data.AggBySiteID.2}
\title{Base Waterbase Dataset for HI, MCR analysis.}
\format{
A data frame with 616795 rows and 12 variables:
\describe{
\item{test_id}{test id}
\item{chemical_name}{chemial name}
...
}
}
\source{
\url{https://cfpub.epa.gov/ecotox/}
}
\usage{
Data.AggBySiteID.2
}
\description{
A dataset containing the waterbase data
}
\keyword{datasets}
