% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data.SSD}
\alias{Data.SSD}
\title{SSD data}
\format{
A data frame with 12386 rows and 38 variables:
\describe{
\item{test_id}{test id}
\item{chemical_name}{chemial name}
...
}
}
\source{
"Posthuma et al. (2019): data-raw/data/Copy of etc4373-sup-0002-supmat.csv"
}
\usage{
Data.SSD
}
\description{
A dataset containing the L.Posthuma SSDs
}
\keyword{datasets}
