% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{data_prepare}
\alias{data_prepare}
\title{Function to check sensitivity to Sample and Chemical per year restrictions}
\usage{
data_prepare(NSample.Year = 5, Nchemicals.Year = 10)
}
\arguments{
\item{Nchemicals.Year}{}
}
\value{

}
\description{
Function to check sensitivity to Sample and Chemical per year restrictions
}
