% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prepare.R
\name{data_prepare_2}
\alias{data_prepare_2}
\title{Function to check sensitivity to Sample and Chemical per year restrictions}
\usage{
data_prepare_2(NSample.Year = 5, Nchemicals.Year = 10)
}
\arguments{
\item{Nchemicals.Year}{}
}
\value{

}
\description{
Function to check sensitivity to Sample and Chemical per year restrictions
}
\examples{

- There are 60776 stations in the repaired station dataset with 48563 having complete longitude and latitude information.


- The chemical monitoring data has 3962763 rows and 31 columns. There are 15990 sites within the stations dataset and 13410 samples with longitude and latitude info.



There are 552 unique CAS values and 373 are included in the SSD dataset. Around 64\% of the data entries with CAS number included in the SSD dataset. There are also 65 CAS values with code "EE" included instead of the CAS number.

2. Keep only “water” matrix. In the analysis matrix procedure matrix, only water was kept in the data, “W-DIS” and “W-SPM” water dissolved and suspended particle matter were excluded.

- Keeping only "water" matrix 3469602 entries were left.




3. Remove records labeled as “unreliable”. In the meta data observation status, only 3343206 records of A (normal record) class were kept, 126396 records of U (unreliable) class were removed



4. Keep only stations not labeled as “legacy”. 3343121 records without the meta data statements "NOTE_LEGACY: Measurement has been confirmed by country to be taken from a highly polluted area" were kept in the final dataset



5. Keep only records for freshwaters. 2989996 records with “LW” and “RW” were kept, whereas 332 records with CW (costal water body), 343585 with GW (groundwater), 9214 TW (transitional).


6. Keep only records for micro pollutants (i.e.: phosphate forms, nitrogen forms, DO, conductivity, biotic elements, etc.) with SSD information.




## Data Curation






1. Records with unit information (ug/L and mg/L) were retained.





2. Result units were all changed to ug/L.





3. Records with missing result mean value, maximum value, number of samples, LOQ value, number of samples below LOQ were excluded

- keep entries with complete data for Average and Maximun values.





4. Records with LOQ below 0 were excluded

- Remove unreliable entries based on missing/unreliable LOQs reporting (LOQ <=0); remove data that do not report "resultQualityNumberOfSamplesBelowLOQ";  Convert reported LOQs to ug/L.
- if resultNumberOfSamples > resultQualityNumberOfSamplesBelowLOQ, then consider the value is above LOQ, however when samples with Max <= LOQ to below LOQ, set it to below LOQ.








5. Records with CAS that has matching SSD info were kept.







6. Dataset restricted to the chemicals (346) for which the SSDs were build on at least 3 species and with slope information (X10LogSSDSlope.ug.L..SigmaChronic.NOEC, X10LogSSDSlope.ug.L..SigmaChronic.NOEC)available.


7. Remove stations/entries do not reporting number of samples, keep only stations reporting more than 5 samples per year to report annual aggregates.


8. There are some repeated entries (multiple Averages for the same CAS, station and Year), the maximum is taken as worst-case scenario for the average.



9. only keep stations measuring at least 10 chemicals the same year


}
