% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{driver}
\alias{driver}
\alias{driver5}
\alias{driver10}
\title{Dataset for Driver chemicals.}
\format{
A vector of CAS numbers that is frequency in drivers greater than 5\%.
\describe{
\item{CAS}{test id}
}

An object of class \code{character} of length 8.
}
\source{
\url{https://github.com/Bayer-Group/WaterBase}
}
\usage{
driver5

driver10
}
\description{
A dataset containing the driver data
}
\keyword{datasets}
