% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_LOQ.R
\name{driverhelper}
\alias{driverhelper}
\title{helper function to extract driver data}
\usage{
driverhelper(data, threshold = 0.1, include0 = F)
}
\arguments{
\item{threshold}{filter HI by threshold}

\item{include0}{whether to include 0 HQ values if the CAS is in the top 3 drivers but there are 0 values in some of the samples.}
}
\value{

}
\description{
helper function to extract driver data
}
