% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_LOQ.R
\name{gen_fig3_2component}
\alias{gen_fig3_2component}
\title{Generate plot for two component comparison. Legacy vs. Current or 6 drivers and all other. ...}
\usage{
gen_fig3_2component(driverdata3, nsample3, sumdriver3, plot = "boxplot")
}
\arguments{
\item{sumdriver3}{}
}
\value{

}
\description{
Generate plot for two component comparison. Legacy vs. Current or 6 drivers and all other. ...
}
