% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_LOQ.R
\name{gen_fig3_wrapper}
\alias{gen_fig3_wrapper}
\title{wrapper function to generate figure 3}
\usage{
gen_fig3_wrapper(
  Data.AggBySiteID.2,
  chemclass,
  LOQ.type = c("LOQ.T0"),
  Exclusion.CAS = c("Refined_HQ_Metals_PAH", "Excluded_Metals", "Current_Use"),
  Data.SSD.1,
  threshold = 1,
  include0 = F,
  threshold_current = 1,
  cutoff = 10,
  addTag = T,
  ...
)
}
\arguments{
\item{threshold}{default 1, HI > 1}

\item{include0}{whether to include HI 0?}

\item{threshold_current}{the current use data, if we need use a smaller threshold for HI, default is 0.1}

\item{cutoff}{N measured > cutoff}

\item{...}{}
}
\value{

}
\description{
wrapper function to generate figure 3
}
