% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_LOQ.R
\name{gen_fig4a_wrapper}
\alias{gen_fig4a_wrapper}
\title{Wrapper function to generate figure 4a}
\usage{
gen_fig4a_wrapper(
  Data.AggBySiteID.2,
  chemclass,
  LOQ.type = c("LOQ.T0"),
  Exclusion.CAS = c("Refined_HQ_Metals_PAH", "Excluded_Metals"),
  Data.SSD.1,
  threshold = 1,
  saveplot = FALSE,
  writetable = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Wrapper function to generate figure 4a
}
