% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_LOQ.R
\name{generate_fig1_1}
\alias{generate_fig1_1}
\title{Function to generate one row of figure 1}
\usage{
generate_fig1_1(
  Res,
  cutoff = 0,
  df = data.frame(x = 1e-06, y = 8),
  test = "Chronic",
  used = c("Mean"),
  exclude.cas = NULL,
  chemclass,
  label3B = TRUE,
  ...
)
}
\arguments{
\item{Res}{summarized data based on waterbase db}

\item{cutoff}{threshold for number of measured chemicals}

\item{...}{}
}
\value{

}
\description{
Function to generate one row of figure 1
}
