% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_LOQ.R
\name{generate_fig1_LOQ}
\alias{generate_fig1_LOQ}
\title{Generate the CEFIC-MIAT plot only for LOQ data.}
\usage{
generate_fig1_LOQ(
  Res_LOQ,
  cutoff = 5,
  df = data.frame(x = 1e-06, y = 8),
  test = "Chronic",
  used = c("Mean"),
  exclude.cas = NULL,
  chemclass,
  label3B = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Generate the CEFIC-MIAT plot only for LOQ data.
}
