% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_LOQ.R
\name{generate_fig3}
\alias{generate_fig3}
\title{Generate Annotated Figure 3}
\usage{
generate_fig3(driverdata2, nsample2, LOQ.type = "LOQ.T0", sumdriver2)
}
\arguments{
\item{driverdata2}{driver data}

\item{nsample2}{summarised driver data}

\item{LOQ.type}{default LOQ.T0}
}
\value{
ggplot object
}
\description{
Generate Annotated Figure 3
}
