% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_LOQ.R
\name{generate_fig4a}
\alias{generate_fig4a}
\title{Function to generate figure 4a}
\usage{
generate_fig4a(
  sumCAS,
  Exclusion.CAS = "Refined_HQ_Metals_PAH",
  chemclass,
  facet = "LOQ.type"
)
}
\arguments{
\item{chemclass}{}
}
\value{

}
\description{
Function to generate figure 4a
}
