% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_LOQ.R
\name{generate_fig4b}
\alias{generate_fig4b}
\title{Function to generate figure 4b}
\usage{
generate_fig4b(
  Data.AggBySiteID.2,
  chemclass,
  Data.SSD.1,
  LOQ.type = c("LOQ.T0"),
  driverdata,
  Exclusion.CAS = "Refined_HQ_Metals_PAH"
)
}
\arguments{
\item{Exclusion.CAS}{}
}
\value{

}
\description{
Function to generate figure 4b
}
