% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_LOQ.R
\name{get_group}
\alias{get_group}
\title{CEFIC-MIAT decision tree table (Valloton &Proce 2016)}
\usage{
get_group(HIv, HQ_maxv, MCRv, AF = 1)
}
\arguments{
\item{HIv}{vector of HI}

\item{HQ_maxv}{vector of max HQ}

\item{MCRv}{vector of MCR}

\item{AF}{assessment factor default being 1}
}
\value{
result of classification of I,II, IIIa, IIIb
}
\description{
CEFIC-MIAT decision tree table (Valloton &Proce 2016)
}
