% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_LOQ.R
\name{getCasData}
\alias{getCasData}
\title{Get 1 CAS info w.r.t. LOQ inside the AggBySiteID waterbase data.}
\usage{
getCasData(
  Data.AggBySiteID.2,
  chemclass,
  Data.SSD.1,
  Exclusion.CAS = "Refined_HQ_Metals_PAH"
)
}
\arguments{
\item{Data.AggBySiteID.2}{the WaterBase data}

\item{chemclass}{Chemical class data}

\item{Data.SSD.1}{SSD information}
}
\value{
data frame with Nmeasured detected, LOQ for each CAS
}
\description{
Get 1 CAS info w.r.t. LOQ inside the AggBySiteID waterbase data.
}
