% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{getCEFIC_MIAT}
\alias{getCEFIC_MIAT}
\title{Calculate the CEFIC_MIAT table}
\usage{
getCEFIC_MIAT(
  Data.AggBySiteID.2,
  chemclass,
  LOQ.type = "LOQ.T0",
  Exclusion.CAS = "Excluded_Metals"
)
}
\arguments{
\item{Exclusion.CAS}{}
}
\value{

}
\description{
Calculate the CEFIC_MIAT table
}
