% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_LOQ.R
\name{getclass}
\alias{getclass}
\title{get the class of the chemicals}
\usage{
getclass(Chem.Group, Approved, Priority, Current_Use)
}
\arguments{
\item{Chem.Group}{chemical group}

\item{Approved}{whether being approved}

\item{Priority}{"STC", "WFD", "STC, WFD"}

\item{Current_Use}{Manually identifies}
}
\value{
class being either Metal, Current_Use, or Banned_Listed.
}
\description{
get the class of the chemicals
}
