% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_LOQ.R
\name{getclass_deprecated}
\alias{getclass_deprecated}
\title{get the class of the chemicals}
\usage{
getclass_deprecated(Chem.Group, Approved, Priority, Candidate.SVHC, SVHC)
}
\arguments{
\item{Chem.Group}{chemical group}

\item{Approved}{whether being approved}

\item{Priority}{"STC", "WFD", "STC, WFD"}

\item{SVHC}{}
}
\value{
class being either Metal, Current_Use, or Banned_Listed.
}
\description{
get the class of the chemicals
}
