% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_LOQ.R
\name{getDriverData1}
\alias{getDriverData1}
\title{Get the main drivers for each site year combination}
\usage{
getDriverData1(
  Data.AggBySiteID.2,
  chemclass,
  LOQ.type = c("LOQ.T0", "LOQ.T1"),
  Exclusion.CAS = c("Excluded_None", "Excluded_Metals", "Current_Use", "Metals",
    "Banned_Listed"),
  Data.SSD.1,
  threshold = 0.1,
  include0 = F,
  cutoff = 0,
  exclude.driver = NULL,
  ...
)
}
\arguments{
\item{Data.AggBySiteID.2}{the WaterBase data}

\item{chemclass}{Chemical class data}

\item{LOQ.type}{handling of LOQ data}

\item{Exclusion.CAS}{excluding classes}

\item{Data.SSD.1}{SSD information}

\item{threshold}{take 0.1 as default}
}
\value{
one data frame with each site year having component 1,2,3,4 with CAS and chemical names.
}
\description{
Get the main drivers for each site year combination
}
