% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_LOQ.R
\name{getHQData_Chronic}
\alias{getHQData_Chronic}
\title{Get HQ chronic data for driver data}
\usage{
getHQData_Chronic(
  Data.AggBySiteID.2,
  chemclass,
  Data.SSD.1,
  LOQ.type,
  driverdata,
  Exclusion.CAS = "Excluded_Metals"
)
}
\arguments{
\item{Data.AggBySiteID.2}{the WaterBase data}

\item{chemclass}{Chemical class data}

\item{Data.SSD.1}{SSD information}

\item{driverdata}{driverdata}
}
\value{
data frame with Nmeasured detected, LOQ for each CAS
}
\description{
Get HQ chronic data for driver data
}
