% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_LOQ.R
\name{getNcas}
\alias{getNcas}
\title{calculate the number of unique CAS}
\usage{
getNcas(Exclusion.CAS, chemclass)
}
\arguments{
\item{Exclusion.CAS}{exlustion option}

\item{chemclass}{chemical grouping information data}
}
\value{
number of cas
}
\description{
calculate the number of unique CAS
}
