% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_LOQ.R
\name{getResData}
\alias{getResData}
\title{Get the relevant data w.r.t. Site and Year}
\usage{
getResData(
  Data.AggBySiteID.2,
  chemclass,
  LOQ.type = c("LOQ.T0", "LOQ.T1"),
  Exclusion.CAS = c("Excluded_None", "Excluded_Metals", "Current_Use", "Metals",
    "Banned_Listed"),
  exclude.substances = NULL,
  Data.SSD.1,
  ...
)
}
\arguments{
\item{Data.AggBySiteID.2}{the WaterBase data}

\item{chemclass}{Chemical class data}

\item{LOQ.type}{handling of LOQ data}

\item{Exclusion.CAS}{excluding classes}

\item{Data.SSD.1}{SSD information}
}
\value{
Stacked data frames with LOQ.type and Exclusion.CAS as identifier
}
\description{
Get the relevant data w.r.t. Site and Year
}
