% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_LOQ.R
\name{getResData1}
\alias{getResData1}
\title{Get 1 dataset}
\usage{
getResData1(
  Data.AggBySiteID.2,
  chemclass,
  LOQ.type = c("LOQ.T0", "LOQ.T1"),
  Exclusion.CAS = c("Excluded_None", "Excluded_Metals", "Current_Use", "Metals",
    "Banned_Listed", "Refined_HQ_Metals_PAH"),
  exclude.substances = NULL,
  Data.SSD.1,
  ...
)
}
\arguments{
\item{Data.AggBySiteID.2}{the WaterBase data}

\item{chemclass}{Chemical class data}

\item{LOQ.type}{handling of LOQ data}

\item{Exclusion.CAS}{excluding classes}

\item{exclude.substances}{exclude substances CAS}

\item{Data.SSD.1}{SSD information}
}
\value{
data frame with MCR, HQ, HI, PAF, and other additional informaion
}
\description{
Get 1 dataset
}
