% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_LOQ.R
\name{getResData_LOQ}
\alias{getResData_LOQ}
\title{Title}
\usage{
getResData_LOQ(
  Data.AggBySiteID.2,
  chemclass,
  Exclusion.CAS = c("Excluded_Metals"),
  exclude.substances = NULL,
  Data.SSD.1,
  LOQ.type = "LOQ.T1",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
