% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_LOQ.R
\name{preprocessWaterbase}
\alias{preprocessWaterbase}
\title{select the designed datasets according to LOQ settings and CAS exclusions}
\usage{
preprocessWaterbase(
  Data.AggBySiteID.2,
  chemclass,
  LOQ.type = c("LOQ.T0", "LOQ.T1"),
  Data.SSD.1,
  Exclusion.CAS = c("Excluded_None", "Excluded_Metals", "Excluded_Metals_PAH",
    "Current_Use", "Metals", "Banned_Listed", "Refined_HQ_Metals_PAH"),
  factor.LOQ = 1,
  LOQ = T,
  SSDrefine = T,
  Exclusion = T,
  HU = T,
  PAF = T,
  useAllMax = F,
  useAllMean = F
)
}
\arguments{
\item{Data.AggBySiteID.2}{data stored in the package, after minimum preprocessing}

\item{chemclass}{chemical information data}

\item{LOQ.type}{"LOQ.T0" or "LOQ.T1"}

\item{Data.SSD.1}{SSD information}

\item{Exclusion.CAS}{exclusion}

\item{factor.LOQ}{use 1 or 1/2 as LOQ handling method}

\item{LOQ}{Always use TRUE}

\item{SSDrefine}{whether to use refined SSD information,
if excluded none, then use original waterbase data with original SSD. otherwise, used PAH, metals refinement}

\item{Exclusion}{Keep it as TRUE}

\item{HU}{whether to calculate HU}

\item{PAF}{whether to calculate PAF}
}
\value{

}
\description{
select the designed datasets according to LOQ settings and CAS exclusions
}
