% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_LOQ.R
\name{selectdata}
\alias{selectdata}
\title{Deprecated: select the designed datasets according to LOQ settings and CAS exclusions}
\usage{
selectdata(
  Data.AggBySiteID.2,
  chemclass,
  LOQ.type = c("LOQ.T0", "LOQ.T1"),
  Exclusion.CAS = c("Excluded_None", "Excluded_Metals", "Excluded_Metals_PAH",
    "Current_Use", "Metals", "Banned_Listed", "Refined_HQ_Metals_PAH"),
  factor.LOQ = 1
)
}
\arguments{
\item{Data.AggBySiteID.2}{data stored in the package, after minimum preprocessing}

\item{chemclass}{chemical info}

\item{LOQ.type}{T0 or T1}

\item{Exclusion.CAS}{exclusion CAS choices}
}
\value{

}
\description{
Deprecated: select the designed datasets according to LOQ settings and CAS exclusions
}
