% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_LOQ.R
\name{sep_6driver}
\alias{sep_6driver}
\title{Get 6driver vs all other data}
\usage{
sep_6driver(
  Data.AggBySiteID.2,
  chemclass,
  LOQ.type = c("LOQ.T0"),
  Exclusion.CAS = "Excluded_Metals",
  Data.SSD.1,
  threshold = 0.1,
  include0 = F,
  cutoff = 0,
  drivers = c("2921882", "15687271", "34123596", "330541", "15307865", "58082")
)
}
\arguments{
\item{drivers}{}
}
\value{

}
\description{
Get 6driver vs all other data
}
