% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAS_LOQ.R
\name{sep_legacy}
\alias{sep_legacy}
\title{Get Legacy vs Curent Data}
\usage{
sep_legacy(
  Data.AggBySiteID.2,
  chemclass,
  LOQ.type = c("LOQ.T0", "LOQ.T1"),
  Exclusion.CAS = c("Excluded_None", "Excluded_Metals", "Current_Use", "Metals",
    "Banned_Listed"),
  Data.SSD.1,
  threshold = 0.1,
  include0 = F,
  cutoff = 0,
  ...
)
}
\arguments{
\item{threshold}{threshold for HI}

\item{cutoff}{}
}
\value{

}
\description{
Get Legacy vs Curent Data
}
