% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCV.R
\name{simConfusion}
\alias{simConfusion}
\title{Simulate the confusion matrix under different assumptions}
\usage{
simConfusion(
  cvs = c(0.3, 0.6, 1, 1.4, 3),
  mdrs = c(1.25, 2, 3, 5),
  Nsim = 12000,
  synergy = F,
  reduction = 0.5,
  q = 0.95
)
}
\arguments{
\item{cvs}{vector of cv}

\item{mdrs}{vector of mdr}

\item{Nsim}{number of simulation}

\item{synergy}{logical}

\item{reduction}{reduction factor}
}
\value{
the confusion matrix
}
\description{
Simulate the confusion matrix under different assumptions
}
