% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCV.R
\name{simCVmat}
\alias{simCVmat}
\title{Simulate MDR matrix}
\usage{
simCVmat(
  cvs,
  mean_sim = 1,
  p = c(0.5, 0.5),
  Nsim = 1000,
  M = 100,
  synergy = F,
  reduction = 0.5
)
}
\arguments{
\item{M}{}
}
\value{

}
\description{
Simulate MDR matrix
}
