% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCV.R
\name{simCVn}
\alias{simCVn}
\title{Simulation function}
\usage{
simCVn(
  cvs,
  mean_sim = 1,
  p = c(0.5, 0.5),
  Nsim = 1000,
  synergy = F,
  reduction = 0.5
)
}
\arguments{
\item{cvs}{A vector of CV}

\item{mean_sim}{mean of the mixture components.}

\item{p}{mixing proportions, added up to 1}
}
\value{
a vector of MDRs
}
\description{
Simulation function
}
