% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stations1}
\alias{stations1}
\alias{stations1_Mean}
\alias{stations1_Max}
\title{Dataset for Station information}
\format{
A vector of CAS numbers that is frequency in drivers greater than 5\%.
\describe{
\item{LOQ.type}{LOQ type}
\item{lat}{latitude}
}

An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 44536 rows and 34 columns.
}
\source{
\url{https://github.com/Bayer-Group/WaterBase}
}
\usage{
stations1_Mean

stations1_Max
}
\description{
A dataset containing the station information
}
\keyword{datasets}
